# src/server/api/routes/organizations.router.js
const vesselController = require("../controllers/vessel.controller");
("use strict");

// router setup

const express = require("express");
const router = express.Router({
  mergeParams: true
});
const {
  authorizeUser,
  authorizeOrganization
} = require("../lib/middleware/auth.middleware");

const ROLES = require("../../constants/roles");

// controllers
const organizationController = require("../controllers/organization.controller");
const vesselsController = require("../controllers/vessel.controller");
const userController = require("../controllers/user.controller");
const voyagesController = require("../controllers/voyages.controller");

// /api/organizations/ :POST
router.post(
  "/",
  authorizeUser(ROLES.SUPER_USER, ROLES.ADMIN),
  (req, res, next) => {
    organizationController
      .createOrganization({
        body: req.body
      })
      .then(result =>
        res.json({
          message: `${req.body.name} added as new organization`,
          id: result
        })
      )
      .catch(next);
  }
);

router.get("/", authorizeUser(ROLES.SUPER_USER), (req, res, next) => {
  organizationController
    .getOrganizations({
      body: req.body
    })
    .then(result => res.json(result))
    .catch(next);
});

// ENDPOINT: /api/organizations/:id :GET
router.get(
  "/:id",
  authorizeUser(ROLES.SUPER_USER, ROLES.ADMIN, ROLES.OPERATOR),
  (req, res, next) => {
    organizationController
      .getOrganizationsById(req.params.id)
      .then(result => res.json(result))
      .catch(next);
  }
);

// ENDPOINT: /api/organizations/:id :PATCH
router.patch(
  "/:id",
  authorizeUser(ROLES.SUPER_USER, ROLES.ADMIN),
  (req, res, next) => {
    organizationController
      .updateOrganizationById({
        body: req.body,
        id: req.params.id
      })
      .then(result => res.json(result))
      .catch(next);
  }
);
//ENDPOINT: /api/organizations/organizations_id/vessels
router.get(
  "/:id/vessels",
  authorizeUser(ROLES.SUPER_USER, ROLES.ADMIN, ROLES.OPERATOR),
  (req, res, next) => {
    vesselsController
      .getVesselsByOrganizationId(req.params.id)
      .then(result => res.json(result))
      .catch(next);
  }
);

//ENDPOINT: /api/organization/:organization_id/voyages?status=ongoing
router.get(
  "/:id/voyages?",
  authorizeUser(ROLES.SUPER_USER, ROLES.ADMIN, ROLES.OPERATOR),
  (req, res, next) => {
    voyagesController
      .getVoyagesByOrganization(req.params.id, req.query.status)
      .then(result => res.json(result))
      .catch(next);
  }
);

//ENDPOINT /api/organizations/:organization_id/users/ :POST
router.post(
  "/:id/users/",
  authorizeUser(ROLES.SUPER_USER),
  authorizeOrganization(),
  (req, res, next) => {
    userController
      .createAdminBySuperUser(req.params.id, req.body)
      .then(result => res.json(result))
      .catch(next);
  }
);

//ENDPOINT /api/organizations/:organization_id/users/ :get
router.get(
  "/:organization_id/users",
  authorizeUser(ROLES.SUPER_USER),
  (req, res, next) => {
    userController
      .getUsersbyOrganizationId(req.params.organization_id)
      .then(result => res.json(result))
      .catch(next);
  }
);

module.exports = router;