# src/server/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Optimum Voyage App",
    "description": "API documentation for the Optimum Voyage app.",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "description": "",
      "name": "Authorization",
      "in": "header"
    }
  },
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/organizations/{organization_id}/users": {
      "post": {
        "description": "Allows creating a user on a specific organization. Currently this is used by super users to create admin users and, thus, this request will only create admin users.\n\nAuthentication: `superuser` only.",
        "summary": "Create Admin User",
        "tags": ["User"],
        "operationId": "OrganizationsAdminByOrganizationIdPost",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateAdminUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/organizations": {
      "post": {
        "description": "Creates an organization. Pass in the name of the organization and a logo for white-labelling the platform.\n\nAuthentication: `superuser` only.",
        "summary": "Create organization",
        "tags": ["Organization"],
        "operationId": "OrganizationsPost",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateorganizationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Get all organizations.\n\nAuthentication: `superuser` only.",
        "summary": "Get organizations",
        "tags": ["Organization"],
        "operationId": "OrganizationsGet",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "description": "Get a given organization by ID.\n\nAuthentication: `superuser`, `admin` & `operator`.",
        "summary": "Get organization by ID",
        "tags": ["Organization"],
        "operationId": "OrganizationsByIdGet",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/organizations/{organization_id}/vessels": {
      "get": {
        "description": "Returns all vessels for a given organization (`admin` & `operators` should only have the option to display vessels for their own organization).\n\nAuthentication: `superuser`, `admin` & `operator`.",
        "summary": "Get vessels by organization",
        "tags": ["Vessel"],
        "operationId": "OrganizationsVesselsByOrganizationIdGet",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/organizations/{organization_id}/voyages?status={voyage_status}": {
      "get": {
        "description": "Returns all voyages for a given organization.\n\nAuthentication: `superuser`, `admin` & `operator`.",
        "summary": "Get voyages by organization",
        "tags": ["Voyage"],
        "operationId": "OrganizationsVoyagesByOrganizationGet",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "voyage_status",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateAdminUserRequest": {
      "title": "CreateAdminUserRequest",
      "example": {
        "name": "John Doe",
        "email": "test3@test3.com",
        "password": "test3",
        "organization_id": 1
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "organization_id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": ["name", "email", "password", "organization_id"]
    },
...
    "CreateorganizationRequest": {
      "title": "CreateorganizationRequest",
      "example": {
        "name": "Teslasss",
        "logo": "/tesla-logo.png"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "logo": {
          "type": "string"
        }
      },
      "required": ["name", "logo"]
    },
...
  },
  "security": [
    {
      "Authorization": []
    }
  ],
...
}